+++ flowserv/model/parameter/files.py
@@ -16,6 +16,7 @@ import os
 from flowserv.model.parameter.base import ParameterBase
 
 import flowserv.error as err
+import flowserv.model.parameter.base as pd
 import flowserv.util as util
 
 
@@ -89,26 +90,21 @@ class FileParameter(ParameterBase):
             try:
                 util.validate_doc(
                     doc,
-                    mandatory=['id', 'type', 'name', 'index', 'isRequired'],
-                    optional=[
-                        'description',
-                        'defaultValue',
-                        'module',
-                        'target'
-                    ]
+                    mandatory=[pd.ID, pd.TYPE, pd.NAME, pd.INDEX, pd.REQUIRED],
+                    optional=[pd.DESC, pd.DEFAULT, pd.MODULE, 'target']
                 )
             except ValueError as ex:
                 raise err.InvalidParameterError(str(ex))
-            if doc['type'] != PARA_FILE:
-                raise ValueError("invalid type '{}'".format(doc['type']))
+            if doc[pd.TYPE] != PARA_FILE:
+                raise ValueError("invalid type '{}'".format(doc[pd.TYPE]))
         return cls(
-            para_id=doc['id'],
-            name=doc['name'],
-            index=doc['index'],
-            description=doc.get('description'),
-            default_value=doc.get('defaultValue'),
-            is_required=doc['isRequired'],
-            module_id=doc.get('module'),
+            para_id=doc[pd.ID],
+            name=doc[pd.NAME],
+            index=doc[pd.INDEX],
+            description=doc.get(pd.DESC),
+            default_value=doc.get(pd.DEFAULT),
+            is_required=doc[pd.REQUIRED],
+            module_id=doc.get(pd.MODULE),
             target=doc.get('target')
         )
 
@@ -146,7 +142,10 @@ class FileParameter(ParameterBase):
                 target = self.default_value
             else:
                 raise err.InvalidArgumentError('missing target path')
-        return InputFile(source=value, target=target, exists=exists)
+        try:
+            return InputFile(source=value, target=target, exists=exists)
+        except TypeError as ex:
+            raise err.InvalidArgumentError(str(ex))
 
     def to_dict(self):
         """Get dictionary serialization for the parameter declaration. Adds