+++ tests/model/parameter/test_numeric_parameter.py
@@ -94,14 +94,14 @@ def test_create_numeric_parameter_error():
     with pytest.raises(KeyError):
         NumericParameter.from_dict({
             'id': '0000',
-            'type': PARA_FLOAT,
+            'dtype': PARA_FLOAT,
             'name': 'X',
             'isRequired': False
         }, validate=False)
     with pytest.raises(err.InvalidParameterError):
         NumericParameter.from_dict({
             'id': '0000',
-            'type': PARA_FLOAT,
+            'dtype': PARA_FLOAT,
             'name': 'X',
             'isRequired': False,
             'range': '0-1'
@@ -109,7 +109,7 @@ def test_create_numeric_parameter_error():
     with pytest.raises(ValueError):
         NumericParameter.from_dict({
             'id': '0000',
-            'type': 'string',
+            'dtype': 'string',
             'name': 'X',
             'index': 0,
             'isRequired': False
@@ -118,7 +118,7 @@ def test_create_numeric_parameter_error():
     with pytest.raises(err.InvalidParameterError):
         NumericParameter.from_dict({
             'id': '0000',
-            'type': PARA_INT,
+            'dtype': PARA_INT,
             'index': 0,
             'isRequired': False
         })
@@ -137,7 +137,7 @@ def test_numeric_parameter(type_id, range):
     """Test creating numeric parameters from dictinaries."""
     doc = {
         'id': '0000',
-        'type': type_id,
+        'dtype': type_id,
         'name': 'X',
         'index': 0,
         'isRequired': False