+++ changelog.md
@@ -34,6 +34,7 @@
 
 ### 0.3.0 - 2020-08-25
 
-* Add manifest file option as parameter for test workflow runs in Jupyter.
 * Option to add specification of output file properties for display purposes in workflow specification.
+* Rename data type element in parameter declarations and result schema columns to 'dtype'.
+* Add manifest file option as parameter for test workflow runs in Jupyter.
 * Test workflow runs using Docker engine.